feat(scheduler): add shard disk repair manager
1. finish shard task process and shard disk repair

with: #22452167

+++ blobstore/scheduler/manual_migrater.go
@@ -60,7 +60,7 @@ func (mgr *ManualMigrateMgr) AddManualTask(ctx context.Context, vuid proto.Vuid,
 	}
 
 	task := &proto.MigrateTask{
-		TaskID:                  client.GenMigrateTaskID(proto.TaskTypeManualMigrate, disk.DiskID, vuid.Vid()),
+		TaskID:                  client.GenMigrateTaskID(proto.TaskTypeManualMigrate, disk.DiskID, uint32(vuid.Vid())),
 		TaskType:                proto.TaskTypeManualMigrate,
 		State:                   proto.MigrateStateInited,
 		SourceIDC:               disk.Idc,
